refactor(gdocs): move custom hooks into own file
